# Use geo-traits for parsing WKT