--- a/Cargo.toml
+++ b/Cargo.toml
@@ -100,6 +100,7 @@ sqlx = { version = "0.7", optional = true, default-features = false, features = 
 ] }
 thiserror = "1"
 tokio = { version = "1", default-features = false, optional = true }
+wkt = "0.11"
 
 
 [dev-dependencies]