--- a/src/io/mod.rs
+++ b/src/io/mod.rs
@@ -27,5 +27,6 @@ pub mod parquet;
 pub mod postgis;
 mod stream;
 pub mod wkb;
+pub mod wkt;
 
 pub use stream::RecordBatchReader;