--- a/src/io/wkt/mod.rs
+++ b/src/io/wkt/mod.rs
@@ -1,0 +1,1 @@
+pub mod reader;